requirements.txt
django>=5.0
langchain>=0.3
langchain-community>=0.3
langchain-text-splitters>=0.3
langchain-huggingface>=0.1
faiss-cpu>=1.8
rank-bm25>=0.2
python-dotenv>=1.0
lxml>=5.0
sentence-transformers>=2.2.2
langdetect>=1.0.9
